+++ CHANGELOG.md
@@ -13,7 +13,7 @@
  - `@ArtemisManager`, like `@ArtemisSystem` but for Managers.
  - `agrotera:matrix` renders _Component Dependency Matrix_ - showing
    mappings between systems, managers and components.
- - `@ArtemisManager` handles `Aspect#one`.
+ - `@ArtemisSystem` handles `Aspect#one`.
  - Less bloat in pom when working with Eclipse + Maven.
 
 ## Version 0.1.1 - 2013-05-05